Begin adding Apache implementation for HTTP POJA
