+++ gradle/libs.versions.toml
@@ -12,6 +12,8 @@ undertow = '2.3.15.Final'
 tomcat = '10.1.26'
 bcpkix = "1.70"
 
+managed-apache-http-core5 = "5.2.5"
+managed-apache-http-client5 = "5.3.1"
 managed-jetty = '11.0.22'
 
 micronaut-reactor = "3.5.0"
@@ -22,8 +24,6 @@ micronaut-validation = "4.7.0"
 google-cloud-functions = '1.1.0'
 kotlin = "1.9.25"
 micronaut-logging = "1.3.0"
-rawhttp-core = "2.4.1"
-rawhttp-cookies = "0.2.1"
 
 # Micronaut
 micronaut-gradle-plugin = "4.4.2"
@@ -48,6 +48,8 @@ junit-platform-engine = { module = "org.junit.platform:junit-platform-suite-engi
 kotlin-stdlib-jdk8 = { module = 'org.jetbrains.kotlin:kotlin-stdlib-jdk8' }
 kotlin-reflect = { module = 'org.jetbrains.kotlin:kotlin-reflect' }
 
+apache-http-core5 = { module = 'org.apache.httpcomponents.core5:httpcore5', version.ref = 'managed-apache-http-core5' }
+apache-http-client5 = { module = 'org.apache.httpcomponents.client5:httpclient5', version.ref = 'managed-apache-http-client5' }
 tomcat-embed-core = { module = 'org.apache.tomcat.embed:tomcat-embed-core', version.ref = 'tomcat' }
 undertow-servlet = { module = 'io.undertow:undertow-servlet', version.ref = 'undertow' }
 jetty-servlet = { module = 'org.eclipse.jetty:jetty-servlet', version.ref = 'managed-jetty' }
@@ -56,9 +58,6 @@ jetty-alpn-server = { module = 'org.eclipse.jetty:jetty-alpn-server', version.re
 jetty-alpn-conscrypt-server = { module = 'org.eclipse.jetty:jetty-alpn-conscrypt-server', version.ref = 'managed-jetty' }
 kotest-runner = { module = 'io.kotest:kotest-runner-junit5', version.ref = 'kotest-runner' }
 bcpkix = { module = "org.bouncycastle:bcpkix-jdk15on", version.ref = "bcpkix" }
-
-rawhttp-core = { module = "com.athaydes.rawhttp:rawhttp-core", version.ref = "rawhttp-core" }
-rawhttp-cookies = { module = "com.athaydes.rawhttp:rawhttp-cookies", version.ref = "rawhttp-cookies" }
 
 google-cloud-functions = { module = 'com.google.cloud.functions:functions-framework-api', version.ref = 'google-cloud-functions' }
 # Gradle
