--- http-poja-llhttp/build.gradle
@@ -1,35 +1,0 @@
-/*
- * Copyright © 2024 Oracle and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-plugins {
-    id("io.micronaut.build.internal.servlet.module")
-}
-
-dependencies {
-    api(projects.micronautHttpPojaCommon)
-    implementation(libs.rawhttp.core)
-    implementation(libs.rawhttp.cookies)
-
-    compileOnly(mn.reactor)
-    compileOnly(mn.micronaut.json.core)
-
-    testImplementation(mnSerde.micronaut.serde.jackson)
-}
-
-micronautBuild {
-    binaryCompatibility {
-        enabled.set(false)
-    }
-}
