+++ settings.gradle
@@ -32,11 +32,11 @@ include 'http-server-jetty'
 include 'http-server-undertow'
 include 'http-server-tomcat'
 include 'http-poja-common'
-include 'http-poja-llhttp'
+include 'http-poja-apache'
 include 'http-poja-test'
 include 'test-suite-http-server-tck-tomcat'
 include 'test-suite-http-server-tck-undertow'
 include 'test-suite-http-server-tck-jetty'
-include 'test-suite-http-server-tck-poja'
+include 'test-suite-http-server-tck-poja-apache'
 include 'test-suite-kotlin-jetty'
 include 'test-sample-poja'
