+++ test-sample-poja/src/main/java/io/micronaut/http/poja/sample/TestController.java
@@ -16,6 +16,7 @@
 package io.micronaut.http.poja.sample;
 
 import io.micronaut.core.annotation.NonNull;
+import io.micronaut.http.HttpRequest;
 import io.micronaut.http.HttpStatus;
 import io.micronaut.http.MediaType;
 import io.micronaut.http.annotation.Controller;
