+++ test-suite-http-server-tck-poja-apache/src/test/resources/META-INF/services/io.micronaut.http.tck.ServerUnderTestProvider
@@ -1,0 +1,1 @@
+io.micronaut.http.server.tck.poja.PojaApacheServerUnderTestProvider
