--- test-suite-http-server-tck-poja/build.gradle
@@ -1,31 +1,0 @@
-plugins {
-    id("io.micronaut.build.internal.servlet.module")
-    id("java-library")
-}
-
-dependencies {
-    testAnnotationProcessor(platform(mn.micronaut.core.bom))
-    testAnnotationProcessor(mn.micronaut.inject.java)
-    testImplementation(platform(mn.micronaut.core.bom))
-    testImplementation(mn.micronaut.inject.java)
-
-    testImplementation(mn.micronaut.http.client)
-    testImplementation(mn.micronaut.http.server.tck)
-    testImplementation(libs.junit.platform.engine)
-    testImplementation(mn.micronaut.jackson.databind)
-    testImplementation(libs.junit.jupiter.engine)
-    testRuntimeOnly(mnLogging.logback.classic)
-
-    testRuntimeOnly(mnValidation.micronaut.validation)
-
-    testImplementation(projects.micronautHttpPojaLlhttp)
-    testImplementation(projects.micronautHttpPojaTest)
-    testImplementation(mnSerde.micronaut.serde.jackson)
-    testImplementation(mn.micronaut.http.client)
-}
-
-micronautBuild {
-    binaryCompatibility {
-        enabled.set(false)
-    }
-}
