--- test-suite-http-server-tck-poja/src/test/java/io/micronaut/http/server/tck/poja/PojaServerUnderTestProvider.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2017-2023 original authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.micronaut.http.server.tck.poja;
-
-import io.micronaut.http.tck.ServerUnderTest;
-import io.micronaut.http.tck.ServerUnderTestProvider;
-
-import java.util.Map;
-
-public class PojaServerUnderTestProvider implements ServerUnderTestProvider {
-
-    @Override
-    public ServerUnderTest getServer(Map<String, Object> properties) {
-        return new PojaServerUnderTest(properties);
-    }
-
-}
