--- test-suite-http-server-tck-poja/src/test/resources/META-INF/services/io.micronaut.http.tck.ServerUnderTestProvider
@@ -1,1 +1,0 @@
-io.micronaut.http.server.tck.poja.PojaServerUnderTestProvider
